Move sample work to separate page
Move sample work content out of slideshow on homepage and onto it's own
page. There was starting to be too much content to comfortably display
one project at a time, so I figure it's probably best to keep it on its
own page, as it is on the current (i.e. soon-to-be-previous) version of
the site.